// src/v3/map/settings/interfaces/IChroma.ts
export interface IChromaSettings {
    /**
     * ## Disable Chroma Events
     * 
     * Disables Chroma for the map.
     * 
     * This will prevent the map from using any Chroma features.
     */
    DisableChromaEvents?: boolean;
    /**
     * ## Disable Environment Enhancements
     * 
     * Disables environment enhancements for the map.
     * 
     * This will prevent the map from using any environment enhancements.
     */
    DisableEnvironmentEnhancements?: boolean;
    /**
     * ## Disable Note Coloring
     * 
     * Disables note coloring for the map.
     * 
     * This will prevent the map from using any note coloring.
     */
    DisableNoteColoring?: boolean;
    /**
     * ## Force Zen Mode Walls
     * 
     * Forces Zen mode walls for the map.
     * 
     * This will force Zen mode walls for the map.
     */
    ForceZenModeWalls?: boolean;
}